+++ src/seekstorm/lib.rs
@@ -4,8 +4,11 @@
 #![doc(html_favicon_url = "http://seekstorm.com/favicon.ico")]
 
 //! # `seekstorm`
+//!
 //! SeekStorm is an open-source, sub-millisecond full-text search library & multi-tenancy server written in Rust.
 //! The **SeekStorm library** can be embedded into your program, while the **SeekStorm server** is a standalone search server to be accessed via HTTP.
+//!
+//!
 //! ### Add required crates to your project
 //! ```text
 //! cargo add seekstorm
@@ -18,19 +21,23 @@
 //! use seekstorm::{index::*,search::*,highlighter::*,commit::Commit};
 //! use tokio::sync::RwLock;
 //! ```
+//!
 //! ### use an asynchronous Rust runtime
 //! ```text
 //! #[tokio::main]
 //! async fn main() -> Result<(), Box<dyn Error + Send + Sync>> {
 //! ```
+//!
 //! ### create index
 //! ```rust
 //! let index_path=Path::new("C:/index/");
+//!
 //! let schema_json = r#"
 //! [{"field":"title","field_type":"Text","stored":false,"indexed":false},
 //! {"field":"body","field_type":"Text","stored":true,"indexed":true},
 //! {"field":"url","field_type":"Text","stored":false,"indexed":false}]"#;
 //! let schema=serde_json::from_str(schema_json).unwrap();
+//!
 //! let meta = IndexMetaObject {
 //! id: 0,
 //! name: "test_index".to_string(),
@@ -38,6 +45,7 @@
 //! tokenizer:TokenizerType::AsciiAlphabetic,
 //! access_type: AccessType::Mmap,
 //! };
+//!
 //! let segment_number_bits1=11;
 //! let serialize_schema=true;
 //! let index=create_index(index_path,meta,&schema,serialize_schema,segment_number_bits1).unwrap();
@@ -55,6 +63,7 @@
 //! {"title":"title2","body":"body2 test","url":"url2"},
 //! {"title":"title3 test","body":"body3 test","url":"url3"}]"#;
 //! let documents_vec=serde_json::from_str(documents_json).unwrap();
+//!
 //! index_arc.index_documents(documents_vec).await;
 //! ```
 //! ### delete documents by document id
@@ -106,6 +115,7 @@
 //!     highlight_markup: true,
 //! },
 //! ];    
+//!
 //! let highlighter=Some(highlighter(highlights, result_object.query_term_strings));
 //! let return_fields_filter= HashSet::new();
 //! let mut index=index_arc.write().await;
@@ -137,7 +147,9 @@
 //! }
 //! ```
 //! ----------------
+//!
 //! ### Faceted search - Quick start
+//!
 //! **Facets are defined in 3 different places:**
 //! 1. the facet fields are defined in schema at create_index,
 //! 2. the facet field values are set in index_document at index time,
@@ -146,6 +158,8 @@
 //!
 //! A minimal working example of faceted indexing & search requires just 60 lines of code. But to puzzle it all together from the documentation alone might be tedious.
 //! This is why we provide a quick start example here:
+//!
+//!
 //! ### Add required crates to your project
 //! ```text
 //! cargo add seekstorm
@@ -158,20 +172,24 @@
 //! use seekstorm::{index::*,search::*,highlighter::*,commit::Commit};
 //! use tokio::sync::RwLock;
 //! ```
+//!
 //! ### use an asynchronous Rust runtime
 //! ```text
 //! #[tokio::main]
 //! async fn main() -> Result<(), Box<dyn Error + Send + Sync>> {
 //! ```
+//!
 //! ### create index
 //! ```rust
 //! let index_path=Path::new("C:/index/");//x
+//!
 //! let schema_json = r#"
 //! [{"field":"title","field_type":"Text","stored":false,"indexed":false},
 //! {"field":"body","field_type":"Text","stored":true,"indexed":true},
 //! {"field":"url","field_type":"Text","stored":true,"indexed":false},
 //! {"field":"town","field_type":"String","stored":false,"indexed":false,"facet":true}]"#;
 //! let schema=serde_json::from_str(schema_json).unwrap();
+//!
 //! let meta = IndexMetaObject {
 //!     id: 0,
 //!     name: "test_index".to_string(),
@@ -179,6 +197,7 @@
 //!     tokenizer:TokenizerType::AsciiAlphabetic,
 //!     access_type: AccessType::Mmap,
 //! };
+//!
 //! let serialize_schema=true;
 //! let segment_number_bits1=11;
 //! let index=create_index(index_path,meta,&schema,serialize_schema,segment_number_bits1,false).unwrap();
@@ -191,6 +210,7 @@
 //! {"title":"title2","body":"body2 test","url":"url2","town":"Warsaw"},
 //! {"title":"title3 test","body":"body3 test","url":"url3","town":"New York"}]"#;
 //! let documents_vec=serde_json::from_str(documents_json).unwrap();
+//!
 //! index_arc.index_documents(documents_vec).await;
 //! ```
 //! ### commit documents
@@ -209,6 +229,7 @@
 //! let query_facets = vec![QueryFacet::String {field: "town".to_string(),prefix: "".to_string(),length: u16::MAX}];
 //! let facet_filter=Vec::new();
 //! //let facet_filter = vec![FacetFilter {field: "town".to_string(),   filter:Filter::String(vec!["Berlin".to_string()])}];
+//!
 //! let result_object = index_arc.search(query, query_type, offset, length, result_type,include_uncommitted,field_filter,query_facets,facet_filter).await;
 //! ```
 //! ### display results
@@ -222,6 +243,7 @@
 //!             highlight_markup: true,
 //!         },
 //!     ];    
+//!
 //! let highlighter2=Some(highlighter(highlights, result_object.query_terms));
 //! let return_fields_filter= HashSet::new();
 //! let index=index_arc.write().await;