+++ src/seekstorm_server/server.rs
@@ -1,10 +1,19 @@
 use base64::{engine::general_purpose, Engine};
+use colored::Colorize;
 use crossbeam_channel::{bounded, select, Receiver};
 use seekstorm::{
     index::{SimilarityType, TokenizerType},
-    ingest::ingest_json,
+    ingest::{ingest_json, ingest_pdf},
 };
-use std::{collections::HashMap, env::current_exe, fs, io, path::Path, sync::Arc};
+use std::{
+    collections::HashMap,
+    env::current_exe,
+    ffi::OsStr,
+    fs::{self, metadata},
+    io,
+    path::Path,
+    sync::Arc,
+};
 use tokio::sync::RwLock;
 
 use crate::{
@@ -118,7 +127,6 @@ pub(crate) async fn initialize(params: HashMap<String, String>) {
 
             recv(receiver_commandline) -> message => {
 
-
                 let m=message.unwrap();
                 let parameter:Vec<&str>=m.split_whitespace().collect();
                 let command=if parameter.is_empty() {""} else {&parameter[0].to_lowercase()};
@@ -128,20 +136,47 @@ pub(crate) async fn initialize(params: HashMap<String, String>) {
                     "ingest" =>
                     {
                         if parameter.len()==1 || parameter.len()==4 {
-                            let apikey_list_clone2=apikey_list_clone.clone();
-                            let apikey_option=if parameter.len()>2 {
-                                get_apikey_hash(parameter[2].to_string(), &apikey_list_clone2).await
+
+                            let data_path_str=if parameter.len()>1 {
+                                parameter[1]
                             } else {
-                                None
+                                "wiki-articles.json"
                             };
+                            let mut data_path = Path::new(&data_path_str);
+                            let mut absolute_path = current_exe().unwrap();
+                            if !data_path.is_absolute() {
+                                absolute_path.pop();
+                                absolute_path.push(data_path_str);
+                                data_path = &absolute_path;
+                            }
 
-                            let mut apikey_list_mut = apikey_list_clone.write().await;
-                            let apikey_object_option=if parameter.len()>2 {
+                            if data_path.exists() {
 
-                                if let Some(apikey_hash) = apikey_option
-                                {
-                                    apikey_list_mut.get_mut(&apikey_hash)
-                                } else if parameter[2]==demo_api_key_base64{
+                                let apikey_list_clone2=apikey_list_clone.clone();
+                                let apikey_option=if parameter.len()>2 {
+                                    get_apikey_hash(parameter[2].to_string(), &apikey_list_clone2).await
+                                } else {
+                                    None
+                                };
+
+                                let mut apikey_list_mut = apikey_list_clone.write().await;
+                                let apikey_object_option=if parameter.len()>2 {
+
+                                    if let Some(apikey_hash) = apikey_option
+                                    {
+                                        apikey_list_mut.get_mut(&apikey_hash)
+                                    } else if parameter[2]==demo_api_key_base64{
+                                        let apikey_quota_object=ApikeyQuotaObject {..Default::default()};
+                                        Some(create_apikey_api(
+                                            &index_path,
+                                            apikey_quota_object,
+                                            &demo_api_key,
+                                            &mut apikey_list_mut,
+                                        ))
+                                    } else{
+                                        None
+                                    }
+                                } else {
                                     let apikey_quota_object=ApikeyQuotaObject {..Default::default()};
                                     Some(create_apikey_api(
                                         &index_path,
@@ -149,69 +184,67 @@ pub(crate) async fn initialize(params: HashMap<String, String>) {
                                         &demo_api_key,
                                         &mut apikey_list_mut,
                                     ))
-                                } else{
-                                    None
-                                }
-                            } else {
-                                let apikey_quota_object=ApikeyQuotaObject {..Default::default()};
-                                Some(create_apikey_api(
-                                    &index_path,
-                                    apikey_quota_object,
-                                    &demo_api_key,
-                                    &mut apikey_list_mut,
-                                ))
-                            };
+                                };
 
-                            if let Some(apikey_object) = apikey_object_option {
+                                if let Some(apikey_object) = apikey_object_option {
 
-                                let index_id=if parameter.len()>3 {
-                                    parameter[3].parse().unwrap_or(0)
-                                } else  if apikey_object.index_list.is_empty() || !apikey_object.index_list.contains_key(&0) {
-                                        let wikipedia_schema_json = r#"
-                                        [{"field":"title","field_type":"Text","stored":true,"indexed":true,"boost":10.0},
-                                        {"field":"body","field_type":"Text","stored":true,"indexed":true},
-                                        {"field":"url","field_type":"Text","stored":true,"indexed":false}]"#;
-                                        create_index_api(
-                                            &index_path,
-                                            "demo_index".to_string(),
-                                            serde_json::from_str(wikipedia_schema_json).unwrap(),
-                                            SimilarityType::Bm25fProximity,
-                                            TokenizerType::UnicodeAlphanumeric,
-                                            apikey_object,
-                                        )
+                                    let index_id=if parameter.len()>3 {
+                                        parameter[3].parse().unwrap_or(0)
+                                    } else if apikey_object.index_list.is_empty() || !apikey_object.index_list.contains_key(&0) {
+                                            let wikipedia_schema_json = r#"
+                                            [{"field":"title","field_type":"Text","stored":true,"indexed":true,"boost":10.0},
+                                            {"field":"body","field_type":"Text","stored":true,"indexed":true},
+                                            {"field":"url","field_type":"Text","stored":true,"indexed":false}]"#;
+                                            create_index_api(
+                                                &index_path,
+                                                "demo_index".to_string(),
+                                                serde_json::from_str(wikipedia_schema_json).unwrap(),
+                                                SimilarityType::Bm25fProximity,
+                                                TokenizerType::UnicodeAlphanumeric,
+                                                apikey_object,
+                                            )
+                                        } else {
+                                            0
+                                        };
+
+                                    if let Some(index_arc) = apikey_object.index_list.get(&index_id) {
+                                        let index_arc_clone=index_arc.clone();
+                                        drop(apikey_list_mut);
+
+                                        let md = metadata(data_path).unwrap();
+                                        if md.is_file() {
+                                            if let Some(extension)=Path::new(&data_path.display().to_string()).extension().and_then(OsStr::to_str) {
+                                                match extension.to_lowercase().as_str() {
+                                                    "pdf" =>{
+                                                        ingest_pdf(index_arc_clone,data_path).await;
+                                                    }
+                                                    "json" =>{
+                                                        ingest_json(index_arc_clone,data_path).await;
+                                                    }
+                                                    _ =>{
+                                                        println!("{} {}","File extension not supported:".red(),extension);
+                                                    }
+                                                }
+                                            } else {
+                                                println!("{} {}","File extension not found:".red(),data_path.display());
+                                            }
+                                        } else {
+                                            ingest_pdf(index_arc_clone,data_path).await;
+                                        }
+
+                                        println!("Set the 'individual API key' in test_api.rest to '{}' when testing the REST API endpoints",if parameter.len()>2 {parameter[2]}else{&demo_api_key_base64});
                                     } else {
-                                        0
-                                    };
-
-                                if let Some(index_arc) = apikey_object.index_list.get(&index_id) {
-                                    let index_arc_clone=index_arc.clone();
-
-                                    drop(apikey_list_mut);
-
-                                    let data_path_str=if parameter.len()>1 {
-                                        parameter[1]
-                                    } else {
-                                        "wiki-articles.json"
-                                    };
-                                    let mut data_path = Path::new(&data_path_str);
-                                    let mut absolute_path = current_exe().unwrap();
-                                    if !data_path.is_absolute() {
-                                        absolute_path.pop();
-                                        absolute_path.push(data_path_str);
-                                        data_path = &absolute_path;
+                                        println!("{} {}","Index not found".red(),index_id);
                                     }
-
-                                    ingest_json(index_arc_clone,data_path).await;
-
-                                    println!("Set the 'individual API key' in test_api.rest to '{}' when testing the REST API endpoints",if parameter.len()>2 {parameter[2]}else{&demo_api_key_base64});
                                 } else {
-                                    println!("Index not found {}",index_id);
+                                    println!("{} {}. Create a valid API key first via REST API or use the demo API key {}","API key not found".red(),parameter[2] ,demo_api_key_base64);
                                 }
                             } else {
-                                println!("API key not found {}. Create a valid API key first via REST API or use the demo API key {}",parameter[2],demo_api_key_base64);
+                                println!("{} {}","Path not found".red(),data_path.display());
                             }
+
                         } else {
-                            println!("Missing parameters: ingest [data_path] [apikey] [index_id]");
+                            println!("{} ingest [data_path] [apikey] [index_id]","Missing parameters:".red());
                         }
                     },
 
@@ -245,6 +278,19 @@ pub(crate) async fn initialize(params: HashMap<String, String>) {
 
                         println!("Server stopped by quit");
                         return;
+                    },
+
+                    "help" =>
+                    {
+                        println!("{}","Server console commands:".yellow());
+                        println!();
+                        println!("{:40} Index wiki-articles.json if present in same directory like seekstorm_server.exe.","ingest".green());
+                        println!("{:40} Index a local data file in JSON, Newline-delimited JSON (ndjson), or Concatenated JSON format.","ingest [data_path] [apikey] [index_id]".green());
+                        println!("{:40} Delete the demo API key and all its indices.","delete".green());
+                        println!("{:40} Stop the server.","quit".green());
+                        println!("{:40} Show this help.","help".green());
+                        println!();
+                        println!("{} https://github.com/SeekStorm/SeekStorm/blob/main/src/seekstorm_server/README.md#console-commands","See details:".yellow());
                     },
 
                     &_ => {